requirements.txt
fastapi
sqlalchemy
requests
uvicorn
pydantic
python-dotenv
python-multipart
python-jose
bcrypt
alembic